+++ collector/livestatus/Collector_test.go
@@ -10,16 +10,16 @@ import (
 
 func TestNewLivestatusCollector(t *testing.T) {
 	livestatus := &MockLivestatus{
-		LivestatusAddress:"localhost:6559",
-		ConnectionType:"tcp",
-		Queries:map[string]string{},
-		isRunning: true,
+		LivestatusAddress: "localhost:6559",
+		ConnectionType:    "tcp",
+		Queries:           map[string]string{},
+		isRunning:         true,
 	}
 	go livestatus.StartMockLivestatus()
 	connector := &Connector{
-		Log:logging.GetLogger(),
-		LivestatusAddress:"localhost:6559",
-		ConnectionType:"tcp",
+		Log:               logging.GetLogger(),
+		LivestatusAddress: "localhost:6559",
+		ConnectionType:    "tcp",
 	}
 	collector := NewLivestatusCollector(make(collector.ResultQueues), connector, "")
 	if collector == nil {
@@ -29,7 +29,7 @@ func TestNewLivestatusCollector(t *testing.T) {
 }
 
 func TestAddTimestampToLivestatusQuery(t *testing.T) {
-	if addTimestampToLivestatusQuery(QueryIcinga2ForNotifications) != fmt.Sprintf(QueryIcinga2ForNotifications, time.Now().Add(intervalToCheckLivestatus / 100 * -150).Unix()) {
+	if addTimestampToLivestatusQuery(QueryIcinga2ForNotifications) != fmt.Sprintf(QueryIcinga2ForNotifications, time.Now().Add(intervalToCheckLivestatus/100*-150).Unix()) {
 		t.Error("addTimestampToLivestatusQuery has changed")
 	}
 }